+++ app/repository/tokens.go
@@ -1,0 +1,7 @@
+package repository
+
+import "gochat/infra/errors"
+
+type ITokens interface {
+	GenerateRTCToken(chanelName, tokenType, uid, role string, expiresIn uint32) (interface{}, *errors.RestErr)
+}