+++ app/serializers/tokens.go
@@ -1,0 +1,9 @@
+package serializers
+
+type TokenReq struct {
+	ChannelName string `json:"channel_name"`
+	TokenType   string `json:"token_type"`
+	UID         string `json:"uid"`
+	Role        string `json:"role"`
+	ExpireIn    uint32 `json:"expire_in,omitempty"`
+}