+++ app/svc/tokens.go
@@ -1,0 +1,10 @@
+package svc
+
+import (
+	"gochat/app/serializers"
+	"gochat/infra/errors"
+)
+
+type ITokens interface {
+	CreateToken(serializers.TokenReq) (interface{}, *errors.RestErr)
+}